aider: Add the page title to the metadata under the 'title' key.
# Aider chat conversation:

USER: Can you add the page title as well in metadata under the 'title' key ?

ASSISTANT: To add the page title in the metadata under the 'title' key, we will modify the `planA_scrapper_danswer.py` file.

Here is the *SEARCH/REPLACE block* for the change:

```python
planA_scrapper_danswer.py
<<<<<<< SEARCH
                parsed_html = web_html_cleanup(soup, self.mintlify_cleanup)

                metadata_content = soup.find_all("div", class_="academy-tag-passive w-dyn-item")
                metadata = {}
                if metadata_content:
                    metadata = {"content": [tag.text for tag in metadata_content]}
=======
                parsed_html = web_html_cleanup(soup, self.mintlify_cleanup)

                metadata_content = soup.find_all("div", class_="academy-tag-passive w-dyn-item")
                metadata = {"title": parsed_html.title or ""}
                if metadata_content:
                    metadata["content"] = [tag.text for tag in metadata_content]
>>>>>>> REPLACE
```

This code will add the page title to the metadata under the 'title' key. If the title is not found, it will be set to an empty string.